requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
flask_bcrypt==0.7.1
marshmallow==3.9.1
requests==2.25.0
flask-migrate==2.5.3
flask-script==2.0.6
python-dotenv==0.15.0
pip~=20.2.4
SQLAlchemy~=1.3.20
bcrypt==3.2.0
py~=1.9.0
Werkzeug==0.16.0
BeautifulSoup4==4.9.3